micro-batching/requirements.txt
faker==28.1.0
apache-airflow==2.10.0
apache-airflow-providers-microsoft-azure==10.4.0
apache-airflow-providers-postgres==5.12.0
apache-airflow-providers-apache-kafka==1.6.0
# apache-airflow-providers-apache-spark==4.10.0
azure-storage-blob==12.22.0
python-dotenv==1.0.1
kafka-python==2.0.2
pyspark==3.5.2